requirements.txt
# Django and framework dependencies
Django>=5.0
djangorestframework>=3.14
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Database
psycopg2-binary>=2.9

# Environment variables
python-dotenv>=1.0

# Image handling
Pillow>=10.0

# Production server
gunicorn>=21.0

# Whitenoise for static files
whitenoise>=6.0

# Django debug toolbar (optional, for development)
django-debug-toolbar>=4.0
